Move update implementation to internals subfolder and renames file to match method name in snake_case
