+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/helpers/index.ts
@@ -27,6 +27,9 @@ export {
   type PreflightDocResult,
   type PreflightNSParams,
   type PreflightNSResult,
+  type DocumentToSave,
+  type ExpectedBulkGetResult,
+  type PreflightGetDocsForBulkUpdateParams,
 } from './preflight_check';
 
 export interface RepositoryHelpers {
