+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/helpers/preflight_check.ts
@@ -5,7 +5,7 @@
  * in compliance with, at your election, the Elastic License 2.0 or the Server
  * Side Public License, v 1.
  */
-
+import { Payload } from '@hapi/boom';
 import type { PublicMethodsOf } from '@kbn/utility-types';
 import { isNotFoundFromUnsupportedServer } from '@kbn/core-elasticsearch-server-internal';
 import type {
@@ -18,6 +18,7 @@ import type { RepositoryEsClient } from '../../repository_es_client';
 import type { PreflightCheckForBulkDeleteParams } from '../internals/repository_bulk_delete_internal_types';
 import type { CreatePointInTimeFinderFn } from '../../point_in_time_finder';
 import {
+  type Either,
   getSavedObjectNamespaces,
   isRight,
   rawDocExistsInNamespaces,
@@ -247,6 +248,73 @@ export class PreflightCheckHelper {
     }
     // any other error from this check does not matter
   }
+
+  /**
+  Pre-flight check fetching all documents for bulkUpdate, regardless of its namespace type */
+  // public async preflightGetDocsForBulkUpdate(params: PreflightGetDocsForBulkUpdate)
+  /**
+   * Fetch multiple saved objects
+   * @returns MgetResponse
+   * @internal
+   */
+  public async preflightGetDocsForBulkUpdate({
+    validObjects,
+    namespace,
+  }: PreflightGetDocsForBulkUpdateParams) {
+    // const { expectedBulkGetResults, namespace } = params;
+    const validObjectsAsRight = validObjects as ExpectedBulkGetResultRight[];
+
+    // if (validObjects.length === 0) {
+    //   // We only have error results; return early to avoid potentially trying authZ checks for 0 types which would result in an exception.
+    //   return {
+    //     // Technically the returned array should only contain SavedObject results, but for errors this is not true (we cast to 'any' below)
+    //     saved_objects: expectedBulkGetResults.map<SavedObject>(
+    //       ({ value }) => value as unknown as SavedObject
+    //     ),
+    //   };
+    // }
+
+    // `objectNamespace` is a namespace string, while `namespace` is a namespace ID.
+    // each of the validObjects in the map might have it's own objectNamespace, we get that using a custom function
+
+    // @TINA note: only using type, id and namespace to get the docs, not searching by attributes
+    const bulkGetDocs = validObjectsAsRight
+      .filter(({ value }) => value.esRequestIndex !== undefined)
+      .map(({ value: { type, id, objectNamespace } }) => ({
+        _id: this.serializer.generateRawId(
+          this.getNamespaceId(objectNamespace, namespace),
+          type,
+          id
+        ),
+        _index: this.getIndexForType(type), // the index in which to get the object
+        _source: ['type', 'namespaces'],
+      }));
+    // @TINA note: initial call to fetch all docs, seems to be issued for single and multinamespace types
+    const bulkGetResponse = bulkGetDocs.length
+      ? await this.client.mget({ body: { docs: bulkGetDocs } }, { ignore: [404], meta: true })
+      : undefined;
+    // fail fast if we can't verify a 404 response is from Elasticsearch
+    if (
+      bulkGetResponse &&
+      isNotFoundFromUnsupportedServer({
+        statusCode: bulkGetResponse.statusCode,
+        headers: bulkGetResponse.headers,
+      })
+    ) {
+      throw SavedObjectsErrorHelpers.createGenericNotFoundEsUnavailableError();
+    }
+    return bulkGetResponse;
+  }
+  /**
+  Private method to get the objects' namespace id for bulkUpdate
+  /**
+   * @returns string | undefined
+   * @internal
+   */
+  private getNamespaceId = (objectNamespace?: string, namespace?: string) =>
+    objectNamespace !== undefined
+      ? SavedObjectsUtils.namespaceStringToId(objectNamespace)
+      : namespace;
 }
 
 /**
@@ -336,4 +404,44 @@ export interface PreflightDocResult {
   checkDocFound: 'not_found' | 'found';
   /** The source of the raw document, if the object already exists in the server's version (unsafe to use) */
   rawDocSource?: GetResponseFound<SavedObjectsRawDocSource>;
+}
+/**
+  @internal
+
+   */
+export type DocumentToSave = Record<string, unknown>;
+/**
+  @internal
+
+   */
+export type ExpectedBulkGetResult = Either<
+  { type: string; id: string; error: Payload },
+  {
+    type: string;
+    id: string;
+    version?: string;
+    documentToSave: DocumentToSave;
+    objectNamespace?: string;
+    esRequestIndex?: number;
+  }
+>;
+/**
+  @internal
+
+   */
+export interface PreflightGetDocsForBulkUpdateParams {
+  validObjects: ExpectedBulkGetResult[];
+  namespace?: string;
+}
+
+interface ExpectedBulkGetResultRight {
+  tag: 'Right';
+  value: {
+    type: string;
+    id: string;
+    version?: string;
+    documentToSave: DocumentToSave;
+    objectNamespace?: string;
+    esRequestIndex?: number;
+  };
 }
