+++ packages/core/saved-objects/core-saved-objects-api-server-internal/src/lib/apis/internals/index.ts
@@ -13,4 +13,4 @@ export {
   type InternalBulkResolveParams,
   type InternalSavedObjectsBulkResolveResponse,
 } from './internal_bulk_resolve';
-export { executeUpdateRetryOnConflict, type PerformUpdateParams } from './update_retry';
+export { executeUpdate, type ExecuteUpdateParams } from './execute_update';
