+++ packages/core/saved-objects/core-saved-objects-api-server/src/apis/bulk_update.ts
@@ -15,7 +15,7 @@ import type { SavedObjectsUpdateOptions, SavedObjectsUpdateResponse } from './up
  * @public
  */
 export interface SavedObjectsBulkUpdateObject<T = unknown>
-  extends Pick<SavedObjectsUpdateOptions<T>, 'version' | 'references'> {
+  extends Omit<SavedObjectsUpdateOptions<T>, 'refresh'> {
   /** The ID of this Saved Object, guaranteed to be unique for all objects of the same `type` */
   id: string;
   /**  The type of this Saved Object. Each plugin can define it's own custom Saved Object types. */
